AI and stats day
mostly worked with the new state machines and AI and transferred most of the behavior from old enemies into new ones and now have things back to how they were before the switch, but new and improved!

I also added a statcounter and got most of the counts done on different actions in the game.  I just have to get it organized to the right format for going into my UI, which I will figure out tomorrow.